Initial html semantic
Initial html semantic wireframe for the home page and added some content.

Included a title for the webpage that includes my name. Added the
four pairs of HTML tags to defne the <header>, <nav>, <div id="main">, and <footer> sections
according to the wireframe. Added a <!--end main> comment immediately after the closing </div>
tag to better identify the end of the main content.

Added some comments as placeholders for future content.

Below the DOCTYPE declaration, I inserted a comment with the name of the document, my first and last names, and the current date.